requirements.txt
Django==4.0.3
djangorestframework==3.13.1
opencv-python==4.5.2.54
python-decouple==3.6
gunicorn==19.9.0
django-utils-six==2.0
django-cors-headers==3.11.0
numpy==1.22.3
tensorflow==2.8.0